[grsec] Applied patch 2.2.0-201007030958
This is grsecurity-2.2.0-2.6.34-201007030958.patch

+++ drivers/gpu/drm/radeon/radeon_display.c
@@ -560,7 +560,7 @@ static void radeon_compute_pll_legacy(struct radeon_pll *pll,
 
 					if (pll->flags & RADEON_PLL_PREFER_CLOSEST_LOWER) {
 						error = freq - current_freq;
-						error = error < 0 ? 0xffffffff : error;
+						error = (int32_t)error < 0 ? 0xffffffff : error;
 					} else
 						error = abs(current_freq - freq);
 					vco_diff = abs(vco - best_vco);
